+++ tests/unittests/test_nodes/test_host_nodes/test_img_detections_filter_node.py
@@ -146,7 +146,7 @@ def test_processing(
     filter: ImgDetectionsFilter,
     request: FixtureRequest,
     img_detections_type: str,
-    duration: int,
+    duration: float,
 ):
     dets: Union[ImgDetectionsExtended, dai.ImgDetections] = request.getfixturevalue(
         img_detections_type
